% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Abrasion}
\alias{Abrasion}
\title{Abrasion loss}
\format{
A data frame with 30 observations on the following 3 variables

\describe{
  \item{\code{perdita}}{ weight loss (in grams per hour)}
  \item{\code{D}}{ hardness (in degrees Shore)}
  \item{\code{Re}}{ tensile strength (in kg/cm\eqn{^2})}
}
}
\source{
Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J., Ostrowski, E. (1994).
  \emph{Small Data Sets}. London Chapman and Hall/CRC.
}
\usage{
Abrasion
}
\description{
Data on the weight loss due to abrasion,
hardness and tensile strength for 30 samples of rubber.
}
\keyword{datasets}
