% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Aids}
\alias{Aids}
\title{Aids mortality}
\format{
Data frame with 14 observations on the following 2 variables

\describe{
  \item{\code{cases}}{ number of deaths}
  \item{\code{periodo}}{ number of period}
}
}
\source{
Dobson, A.J. (1990). \emph{An Introduction to Generalized Linear Models}. London: CRC Press.
}
\usage{
Aids
}
\description{
Number of AIDS deaths in a sequence of three-months periods between 1983 and 1986.
}
\keyword{datasets}
