% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Alligators}
\alias{Alligators}
\title{Alligator food choice data}
\format{
A data frame with 40 rows and 4 variables:

\describe{

\item{\code{foodchoice}}{primary food type, in volume, found in an alligator’s stomach, with levels \code{fish}, \code{invertebrate}, \code{reptile}, \code{bird}, \code{other}}

\item{\code{lake}}{lake of capture with levels \code{Hancock}, \code{Oklawaha}, \code{Trafford}, \code{George}}

\item{\code{size}}{size of the alligator with levels \code{<=2.3} meters long and \code{>2.3} meters long}

\item{\code{Freq}}{number of alligators for each foodchoice, lake, gender and size combination}

}
}
\source{
The alligators data set is analysed in Agresti (2002, Subsection 7.1.2).

This is an edited version of the original data set, which is available at
\url{http://www.stat.ufl.edu/~aa/glm/data/}
}
\usage{
Alligators
}
\description{
Alligator food choice data
}
\references{
Agresti, A. (2002). \emph{Categorical Data Analysis}. New York: Wiley.
}
\keyword{datasets}
