% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ants}
\alias{Ants}
\title{Ants and sandwiches}
\format{
Data frame with 48 observations on the following 5 variables

\describe{

\item{\code{Bread}}{integer indicator for the kind of bread (1=rye, 2=wheatmeal, 3=multigrain, 4=white)}

\item{\code{Filling}}{integer indicator for the kind of filling (1=vegemite, 2=peanut butter, 3=ham and pickles)}

\item{\code{Butter}}{indicator for butter (1=butter, -1=no butter)}

\item{\code{Ant_count}}{number of captured ants}

\item{\code{Order}}{order of the experiment}
}
}
\source{
Mackisack, M. (2017). What is the use of experiments conducted by Statistics students?
\emph{Journal of Statistics Education}, \bold{2}, 12-15.
}
\usage{
Ants
}
\description{
The dataset refers to an experiment
carried out by some students of an Australian university.
}
\keyword{datasets}
