% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Aziende}
\alias{Aziende}
\title{Number of closed businesses}
\format{
Data frame with 16 observations on the following 4 variables

\describe{

\item{\code{regione}}{integer indicator for the region}

\item{\code{numero}}{number of closed businesses}

\item{\code{dimensione}}{average dimension of the businesses}

\item{\code{salario}}{average individual salary}
}
}
\source{
Salvan, A., Sartori, N., Pace, L. (2020). \emph{Modelli lineari generalizzati}. Milano: Springer-Verlag.
}
\usage{
Aziende
}
\description{
The data refers to the number of business that
have closed their activity in the first trimester of 2005 in 16 Italian regions.
}
\keyword{datasets}
