% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bartlett2}
\alias{Bartlett2}
\title{Bartlett data on plum root cuttings}
\format{
A data frame with 4 rows and 4 columns related to the cross-classification of 960 observations.
The variables are:

\describe{

\item{\code{Alive}}{number of plum root cuttings survived}

\item{\code{Dead}}{number of plum root cuttings dead}

\item{\code{Time}}{factor w/ 2 levels (\code{Now, Spring})}

\item{\code{Length}}{factor w/ 2 levels (\code{Long, Short})}
}
}
\source{
Hand, D. and Daly, F. and Lunn, A. D.and
McConway, K. J. and Ostrowski, E. (1994). \emph{A Handbook of Small Data
Sets}.  London: Chapman & Hall, p. 15, # 19.
}
\usage{
Bartlett2
}
\description{
In an experiment to investigate the effect of cutting length (two levels)
and planting time (two levels) on the survival of plum root cuttings, 240
cuttings were planted for each of the 2 x 2 combinations of these factors,
and their survival was later recorded.
}
\references{
Bartlett, M. S. (1935).  Contingency Table
Interactions \emph{Journal of the Royal Statistical Society}, Supplement,
1935, 2, 248-252.
}
\seealso{
\code{\link{Bartlett}} for the same data in table format
}
\keyword{datasets}
