% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Beetles}
\alias{Beetles}
\title{Deaths of flour beetles}
\format{
A data frame with 8 observations on the following 3 variables
\describe{
  \item{\code{num}}{numbers of beetles exposed}
  \item{\code{uccisi}}{numbers of beetles dying}
  \item{\code{logdose}}{concentration of carbon disulphide (mg. per litre) in logarithmic scale}
}
}
\source{
Bliss, C. I. (1935).The calculation of the dosage-mortality curve. \emph{Annals of Applied BIology},
 \bold{22}, 134-167.
}
\usage{
Beetles
}
\description{
Number of adult flour beetles which died following a 5-hour exposure to gaseous carbon disulphide.
}
\seealso{
\code{\link{Beetles10}} for an ungrouped version of this data
}
\keyword{datasets}
