% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Beetles10}
\alias{Beetles10}
\title{Deaths of flour beetles}
\format{
A data frame with 481 observations on the following 2 variables
\describe{
  \item{\code{log.dose10}}{concentration of carbon disulphide (mg. per litre) in logarithmic scale}
  \item{\code{ucciso}}{indicator variable of death (0: survived, 1: dead)}
}
}
\source{
Bliss, C. I. (1935).The calculation of the dosage-mortality curve. \emph{Annals of Applied BIology},
 \bold{22}, 134-167.
}
\usage{
Beetles10
}
\description{
Survival adult flour beetles which died following a 5-hour exposure to gaseous carbon disulphide.
}
\seealso{
\code{\link{Beetles}} for a grouped version of these data
}
\keyword{datasets}
