% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bioassay}
\alias{Bioassay}
\title{Biological experiment}
\format{
A data frame with 10 observations on the following 3 variables
\describe{
  \item{\code{z}}{dose level}
  \item{\code{den}}{number of exposed}
  \item{\code{y}}{number of observed events}
}
}
\source{
Finney, D.J. (1947). \emph{Probit Analysis}. Cambridge: Cambridge University Press.
}
\usage{
Bioassay
}
\description{
Number of events observed in a biological experiment with different dose exposure.
}
\keyword{datasets}
