% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Britishdoc}
\alias{Britishdoc}
\title{British doctors study}
\format{
A data frame with 10 observations on the following 4 variables
\describe{
  \item{\code{age}}{factor with 5 levels (\code{35-44, 45-54, 55-64, 65-74, 75-84})}
  \item{\code{smoke}}{factor with 2 levels (\code{n, y})}
  \item{\code{person.years}}{total number of observed person-years}
  \item{\code{deaths}}{number of observed deaths by coronary disease}
}
}
\source{
Agresti, A. (2015). \emph{Foundations of Linear and Generalized Linear Models}. Hoboken: Wiley.
}
\usage{
Britishdoc
}
\description{
Study on coronary deaths involving British doctors.
}
\keyword{datasets}
