\name{Calcium}
\alias{Calcium}
\docType{data}
\title{Calcium Uptake Data}
\description{
Data on the uptake of calcium by cells suspended in a
radioactive solution, as a function of time.
}
\format{
  A data frame with 27 observations on the following 2 variables
  \describe{
    \item{\code{time}}{The time (in minutes) that the cells were suspended in the
          solution}
    \item{\code{cal}}{The amount of calcium uptake (nmoles/mg)}
  }
}
\details{
      Howard Grimes from the Botany Department, North Carolina State
     University, conducted an experiment for biochemical analysis of
     intracellular storage and transport of calcium across plasma
     membrane.  Cells were suspended in a solution of radioactive
     calcium for a certain length of time and then the amount of
     radioactive calcium that was absorbed by the cells was measured.
     The experiment was repeated independently with 9 different times
     of suspension each replicated 3 times.
}
\source{
Rawlings, J.O. (1988) \emph{Applied Regression Analysis}.  Wadsworth and
     Brooks/Cole Statistics/Probability Series.

Package \code{SMPracticals}
}
\references{
Davison, A. C. (2003)  \emph{Statistical Models}.  Cambridge University Press.
Page 469.


}

\keyword{datasets}
