% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cement}
\alias{Cement}
\title{Tensile strength of cement}
\format{
An object of class \code{data.frame} with 21 rows and 2 columns.
}
\source{
Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J., Ostrowski, E. (1994).
  \emph{Small Data Sets}. London Chapman and Hall/CRC.
}
\usage{
Cement
}
\description{
Experiment where different batches of cement were tested for tensile strength after different curing times.
}
\details{
\describe{
  \item{\code{tempo}}{curing times (in days)}
  \item{\code{resistenza}}{tensile strength (kg/cm$^2$)}
}
}
\keyword{datasets}
