\name{Chimps}
\alias{Chimps}
\docType{data}
\title{Chimpanzee Learning Data}
\description{
These are the times in minutes taken for four chimpanzees to learn each of four
words.
}
\format{
  A data frame with 40 observations on the following 3 variables
  \describe{
    \item{\code{chimp}}{a factor with levels 1-4}
    \item{\code{word}}{a factor with 1-10}
    \item{\code{y}}{learning time (minutes)}
  }
}
\source{
Brown, B. W. and Hollander, M. (1977) \emph{Statistics: A Biomedical
  Introduction}. New York: Wiley.

Package \code{SMPracticals}
}
\references{
Davison, A. C. (2003)  \emph{Statistical Models}.  Cambridge University Press.
Page 485.


}

\keyword{datasets}
