% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Chlorsulfuron}
\alias{Chlorsulfuron}
\title{Chlorsulfuron Data}
\format{
A data frame with 51 observations on the following 3 variables
\describe{
  \item{\code{gruppo}}{indicator variable for each tested dose}
  \item{\code{dose}}{the tested dose (nmol/l)}
  \item{\code{area}}{the callus area (mm^2)}
}
}
\source{
Package nlreg

 Seiden, P., Kappel, D. e Streibig, J.C. (1998). Response of Brassica napus L. tissue culture to
 metsulfuron methyl and chlorsulfuron. \emph{Weed Research}, \bold{38}, 221-228.
}
\usage{
Chlorsulfuron
}
\description{
Bioassay on the action of the herbicide chlorsulfuron on the callus area of colonies of Brassica napus L.
The experiment consists of 51 measurements for 10 different dose levels.
The design is unbalanced: the number of replicates per dose varies from a minimum of 5 to a maximum of 8.
}
\keyword{datasets}
