% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Clotting}
\alias{Clotting}
\title{Blood clotting times}
\format{
Data frame with 18 observations on the following 3 variables

\describe{
  \item{\code{u}}{plasma concentration (in precentage)}
  \item{\code{tempo}}{clotting time  (in seconds)}
  \item{\code{lotto}}{lot (factor with two levels: uno, due)}
}
}
\source{
McCullagh, P. and Nelder, J. A. (1989) \emph{Generalized Linear
  Models} (2nd Edition).  London: Chapman and Hall.
}
\usage{
Clotting
}
\description{
Mean blood clotting times in seconds for nine percentage concentrations of
normal plasma and two lots of clotting agent.
}
\keyword{datasets}
