% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Customer}
\alias{Customer}
\title{Bus customer satisfaction}
\format{
A data frame with 12231 observations on the following 2 variables
\describe{
  \item{\code{y}}{level of satisfaction, factor with 5 levels (\code{Neutral, Satisfied, Unsatisfied, Very satisfied, Very unsatisfied})}
  \item{\code{delay}}{bus delay (in minutes)}
}
}
\source{
Madsen, H. e Thyregod, P. (2010). \emph{Introduction to General and Generalized Linear Models}. Boca Raton,
  CRC Press.
}
\usage{
Customer
}
\description{
Survey on the customer satisfaction among passengers of a certain bus line.
}
\seealso{
\code{\link{Customer3}} for the same data in table format
}
\keyword{datasets}
