% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Customer3}
\alias{Customer3}
\title{Bus customer satisfaction}
\format{
The data are stored as a frequency table.
Data frame with 4 observations on the following 6 variables
\describe{
  \item{\code{delay}}{bus delay (in minutes)}
  \item{\code{Verydissatisfied}}{frequency of "Very dissatisfied" replies to the survey}
  \item{\code{Dissatisfied}}{frequency of "Dissatisfied" replies to the survey}
  \item{\code{Neutral}}{frequency of "Neutral" replies to the survey}
  \item{\code{Satisfied}}{frequency of "Satisfied" replies to the survey}
  \item{\code{Verysatisfied}}{frequency of "Very satisfied" replies to the survey}
}
}
\source{
Madsen, H. e Thyregod, P. (2010). \emph{Introduction to General and Generalized Linear Models}. Boca Raton,
  CRC Press.
}
\usage{
Customer3
}
\description{
Survey on the customer satisfaction among passengers of a certain bus line.
}
\seealso{
\code{\link{Customer}} for the individual level data
}
\keyword{datasets}
