\name{Dogs}
\alias{Dogs}
\docType{data}
\title{Dogs data}
\description{Measurements of left ventricular volume and parallel conductance volume
on five dogs under eight different load conditions
}
\usage{Dogs}

\format{
  Data frame with 40 observations on the following 4 variables
  \describe{
    \item{\code{dog}}{dog number}
    \item{\code{condition}}{load condition}
    \item{\code{y}}{left ventricular volume}
    \item{\code{x}}{parallel conductance volume}
  }
}

\source{
Package \code{dobson}

Dobson, A. J. and Barnett A. (2008). \emph{An Introduction to Generalized Linear Models}, 3rd ed. Boca Raton: CRC
Press.}
\references{Boltwood, C. M., R. Appleyard, and S. A. Glantz (1989). Left ventricular
volume measurement by conductance catheter in intact dogs: the parallel
conductance volume increases with end-systolic volume. \emph{Circulation} 80,
1360–1377.}
\keyword{datasets}

