% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Drugs}
\alias{Drugs}
\title{Student Substance Use}
\format{
A data frame with 8 observations on the following 4 variables
\describe{
  \item{\code{alc}}{alcohol use, factor with 2 levels (\code{no, yes})}
  \item{\code{sig}}{sigarettes use, factor with 2 levels (\code{no, yes})}
  \item{\code{mar}}{marijuana use, factor with 2 levels (\code{no, yes})}
  \item{\code{count}}{frequency of students in the cross classification of the previous three variables}
}
}
\source{
Agresti, A. (2015). \emph{Foundations of Linear and Generalized Linear Models}. Hoboken: Wiley.
}
\usage{
Drugs
}
\description{
Survey on alcohol, cigarettes, or marijuana use collected on 2276 students in their final year of high school
in a rural area near Dayton, Ohio.
}
\seealso{
\code{\link{Drugs2}} for a different format of the same data and \code{\link{Drugs3}}
for an extended version of the data with additional variables.
}
\keyword{datasets}
