% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Drugs2}
\alias{Drugs2}
\title{Student Substance Use}
\format{
A data frame with 4 observations on the following 5 variables
\describe{
  \item{\code{alc}}{alcohol use, factor with 2 levels (\code{no, yes})}
  \item{\code{sig}}{sigarettes use, factor with 2 levels (\code{no, yes})}
  \item{\code{M_yes}}{frequency of students that have tried marijuana}
  \item{\code{M_no}}{frequency of students that have never tried marijuana}
  \item{\code{n}}{frequency of students in the cross classification of variables \code{alc} and \code{sig}}
}
}
\source{
Agresti, A. (2015). \emph{Foundations of Linear and Generalized Linear Models}. Hoboken: Wiley.
}
\usage{
Drugs2
}
\description{
Survey on alcohol, cigarettes, or marijuana use made on 2276 students in their final year of high school
in a rural area near Dayton, Ohio.
}
\seealso{
\code{\link{Drugs}} for a different format of the same data and \code{\link{Drugs3}}
for an extended version of the data with additional variables.
}
\keyword{datasets}
