% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Drugs3}
\alias{Drugs3}
\title{Student Substance Use}
\format{
A data frame with 32 observations on the following 6 variables
\describe{
  \item{\code{alcohol}}{alcohol use, factor with 2 levels (\code{no, yes})}
  \item{\code{cigarette}}{cigarettes use, factor with 2 levels (\code{no, yes})}
  \item{\code{marijuana}}{marijuana use, factor with 2 levels (\code{no, yes})}
  \item{\code{gender}}{factor with 2 levels (\code{Female, Male})}
  \item{\code{race}}{factor with 2 levels (\code{Other, White})}
  \item{\code{Freq}}{frequency of students in the cross classification of the previous five variables}
}
}
\source{
Agresti, A. (2015). \emph{Foundations of Linear and Generalized Linear Models}. Hoboken: Wiley.
}
\usage{
Drugs3
}
\description{
Survey on alcohol, cigarettes, or marijuana use made on 2276 students in their final year of high school
in a rural area near Dayton, Ohio.
}
\seealso{
\code{\link{Drugs}} and \code{\link{Drugs2}} for a reduced version of this data,
with fewer variables, in two different formats.
}
\keyword{datasets}
