% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Esito}
\alias{Esito}
\title{Recreational activities and university performance}
\format{
A data frame with 18 observations on the following 4 variables
\describe{
  \item{\code{freq}}{frequency of students in the in the cross classification of the following three variables}
  \item{\code{sex}}{factor with 2 levels (\code{f, m})}
  \item{\code{ore}}{weekly hours of recreational activities, factor with 3 levels (\code{m10}, less than 10 hours;
  \code{m15}, between 10 and 15 hours; \code{m20}, more than 15 hours)}
  \item{\code{voto}}{university performance in a given exam, factor with 3 levels (\code{ins}, not sufficient;
  \code{suff}, sufficient; \code{buono}, good)}
}
}
\source{
Salvan, A., Sartori, N., Pace, L. (2020). \emph{Modelli lineari generalizzati}. Milano: Springer-Verlag.
}
\usage{
Esito
}
\description{
Survey on the effect of recreational activities on university performance collected on 485 students.
}
\keyword{datasets}
