% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Germination}
\alias{Germination}
\title{Seed Germination}
\format{
Data frame with 21 observations in the following 4 variables

\describe{

\item{\code{s}}{number of germinated seeds}

\item{\code{m}}{total number of seeds}

\item{\code{seed}}{seed indicator, factor with 2 levels (\code{073, 075})}

\item{\code{root}}{root indicator, factor with 2 levels (\code{C, F})}
}
}
\source{
Cox, D.R. e Snell, E.J. (1989). \emph{Analysis of Binary Data}, 2nd ed. London: Chapman & Hall/CRC.
}
\usage{
Germination
}
\description{
Factorial experiment on the germination of two different kind of seeds (Orobanche aegyptiaca 75 and
 Orobanche aegyptiaca 73) in two different experimental conditions (bean or cucumber root).
}
\keyword{datasets}
