% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Heart}
\alias{Heart}
\title{Creatinine kinase and heart attacks}
\format{
Data frame with 13 observations and the following 4 variables
\describe{
  \item{\code{mck}}{central value of the class of Creatinine kinase level in variable \code{ck}}
  \item{\code{ck}}{class of Creatinine kinase level (in IU per litre), factor with 13 levels
  (\code{Below 40}, \code{40-80}, \ldots, \code{480 and over})}
  \item{\code{ha}}{number of patients with diagnosed heart attack}
  \item{\code{nha}}{number of patients without heart attack}
}
}
\source{
Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J., Ostrowski, E. (1994).
  \emph{Small Data Sets}. London Chapman and Hall/CRC.
}
\usage{
Heart
}
\description{
Data on diagnosed heart attacks in a sample of 360 patients hospitalized with suspected heart attack.
}
\keyword{datasets}
