% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Homicide}
\alias{Homicide}
\title{Homicide data}
\format{
A data frame with 1308 observations on the following 2 variables
\describe{
  \item{\code{race}}{indicator of self-identified race (\code{0}, white; \code{1}, black)}
  \item{\code{count}}{number of known victims of murder in the last year}
}
}
\source{
Agresti, A. (2015). \emph{Foundations of Linear and Generalized Linear Models}. Hoboken: Wiley.

  \url{http://www.stat.ufl.edu/~aa/glm/data}
}
\usage{
Homicide
}
\description{
Survey on number of victims of murder known in the past year by race.
}
\keyword{datasets}
