% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Infant}
\alias{Infant}
\title{Infant survival}
\format{
A data frame with 16 observations on the following 5 variables
\describe{
  \item{\code{survival}}{survival of the infant, factor with 2 levels (\code{No, Yes})}
  \item{\code{gestation}}{length of gestation (in days), factor with 2 levels (\code{<=260, >260})}
  \item{\code{smoking}}{number of cigarettes per day smoked by the mother, factor with 2 levels (\code{<5, >5})}
  \item{\code{age}}{age of the mother (in years), factor with 2 levels (\code{<30, >30})}
  \item{\code{Freq}}{frequency of infant in the cross classification of the previous 4 variables}
}
}
\source{
Agresti, A. (2013). \emph{Categorical Data Analysis}, 3rd ed. New York: Wiley.
}
\usage{
Infant
}
\description{
Study that relates the survival of infants to length of gestation, age and smoking habit of mothers.
}
\keyword{datasets}
