\name{Kyphosis}
\alias{Kyphosis}
\title{Data on Children who have had Corrective Spinal Surgery}
\description{
  Data on children who have had corrective spinal surgery.
}
\usage{
Kyphosis
}
\format{
Data frame with 81 observations on the following 4 variables
\describe{
\item{\code{Kyphosis}}{
a factor with levels
\code{absent}
\code{present}
indicating if a kyphosis (a type of deformation)
was present after the operation.
}
\item{\code{Age}}{
in months
}
\item{\code{Number}}{
the number of vertebrae involved
}
\item{\code{Start}}{
the number of the first (topmost) vertebra operated on.
}
}}
\source{
  Hastie, T.J. and Tibshirani, R.J. (1990). \emph{Generalized Additive Models}. London: Chapman & Hall/CRC.
}
\keyword{datasets}
