% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mental}
\alias{Mental}
\title{Mental impairment}
\format{
Data frame with 40 observations in the following 3 variables

\describe{

\item{\code{menom}}{mental health status on an ordinal scale (\code{1}, well; \code{2}, mild symptom formation;
 \code{3}, moderate symptom formation; \code{4}, impaired)}

\item{\code{sse}}{Socioeconomic status (\code{1}, high; \code{0}, low)}

\item{\code{eventi}}{life events index, a composite measure of the number and severity of important life events
that occurred to the subject within the past 3 years, such as the birth of a child, a new job, a divorce,
or a death in the family}
}
}
\source{
Agresti, A. (2015). \emph{Foundations of Linear and Generalized Linear Models}. Hoboken: Wiley.
}
\usage{
Mental
}
\description{
Study of mental health for a random sample of adult residents of Alachua County, Florida.
}
\keyword{datasets}
