% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLGdata-package.R
\docType{package}
\name{MLGdata-package}
\alias{MLGdata}
\alias{MLGdata-package}
\title{MLGdata: Datasets used in Salvan, Sartori and Pace (2020) Modelli Lineari Generalizzati}
\description{
Datasets used in the book "Modelli Lineari Generalizzati" (Salvan, Sartori and Pace, 2020, Springer-Verlag).
}
\keyword{internal}
