% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Neonati}
\alias{Neonati}
\title{Weight at birth}
\format{
Data frame with 32 observations on the following 3 variables
\describe{
  \item{\code{peso}}{weigth at birth (in grams)}
  \item{\code{durata}}{duration of gestation (in weeks)}
  \item{\code{fumo}}{a factor with levels \code{F} (smoker), \code{NF} (non smoker)}
}
}
\source{
Daniel, W.W. (1999). \emph{Biostatistics: A Foundation for Analysis in the Health Sciences}.
  New York: Wiley.
}
\usage{
Neonati
}
\description{
Data on the weight at birth,
the duration of the gestation, and the smoke habit of the mother for 32 newborns.
}
\keyword{datasets}
