% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ohio}
\alias{Ohio}
\title{Ohio Children Wheeze Status}
\format{
Data frame with 2148 observations on the following 4 variables

\describe{

\item{\code{resp}}{an indicator of wheeze status (1=yes, 0=no)}

\item{\code{id}}{a numeric vector for subject id}

\item{\code{age}}{a numeric vector of age, 0 is 9 years old}

\item{\code{smoke}}{an indicator of maternal smoking at the first
    year of the study}
}
}
\source{
Package \code{geepack}
}
\usage{
Ohio
}
\description{
The dataset is a
subset of the six-city study, a longitudinal study of the health effects of
air pollution.
}
\references{
Fitzmaurice, G.M. and Laird, N.M. (1993) A likelihood-based
    method for analyzing longitudinal binary responses, \emph{Biometrika}
    \bold{80}: 141--151.

    Halekoh, U., Højsgaard, S. e Yan, J. (2005).
    The R package geepack for generalized estimating
   equations. \emph{Journal of Statistical Software}, \bold{15}, 1-11.
}
\keyword{datasets}
