% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Orthodont}
\alias{Orthodont}
\title{Growth curve data on an orthdontic measurement}
\format{
Data frame with 27 observations in the following 5 variables
\describe{
  \item{\code{genere}}{gender of the subject, factor with 2 levels (\code{F, M})}
  \item{\code{dist8a}}{measurement of the orthodontic distance (in mm) at age 8}
  \item{\code{dist10a}}{measurement of the orthodontic distance (in mm) at age 10}
  \item{\code{dist12a}}{measurement of the orthodontic distance (in mm) at age 12}
  \item{\code{dist14a}}{measurement of the orthodontic distance (in mm) at age 14}
}
}
\source{
Pinheiro, J.C. and Bates, D.M. (2000). \emph{Mixed Effects Models in S and S-PLUS}. New York: Springer.

Package \code{nlme}
}
\usage{
Orthodont
}
\description{
Study of the change in an orthdontic measurement over time for 27 young subjects.
}
\references{
Potthoff, R.F. and Roy, S.N. (1964). A generalized multivariate analysis of variance
model useful especially for growth curve problems. \emph{Biometrika}, \bold{51}, 313-326.
}
\seealso{
\code{\link{Orthodont1}} for the same data in an different format
}
\keyword{datasets}
