% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Orthodont1}
\alias{Orthodont1}
\title{Growth curve data on an orthdontic measurement}
\format{
Data frame with 108 observations in the following 4 variables
\describe{
  \item{\code{caso}}{subject index}
  \item{\code{genere}}{gender of the subject, factor with 2 levels (\code{F, M})}
  \item{\code{eta}}{age of the subject}
  \item{\code{y}}{measurement of the orthodontic distance (in mm)}
}
}
\source{
Pinheiro, J.C. and Bates, D.M. (2000). \emph{Mixed Effects Models in S and S-PLUS}. New York: Springer.

Package \code{nlme}
}
\usage{
Orthodont1
}
\description{
Study of the change in an orthdontic measurement over time for 27 young subjects.
}
\references{
Potthoff, R.F. and Roy, S.N. (1964). A generalized multivariate analysis of variance
model useful especially for growth curve problems. \emph{Biometrika}, \bold{51}, 313-326.
}
\seealso{
\code{\link{Orthodont}} for the same data in a different version
}
\keyword{datasets}
