% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Seed}
\alias{Seed}
\title{Seed germination}
\format{
A data frame with 20 observations on the following 2 variables
\describe{
  \item{\code{fert}}{level of fertilizer used}
  \item{\code{x}}{indicator of germination of the seed(\code{1}, yes; \code{0}, no)}
}
}
\source{
Salvan, A., Sartori, N., Pace, L. (2020). \emph{Modelli lineari generalizzati}. Milano: Springer-Verlag.
}
\usage{
Seed
}
\description{
This is an artificial dataset representing an experiment relating probability of
germination of seeds to the level of fertilizer used.
}
\keyword{datasets}
