% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Snore}
\alias{Snore}
\title{Snoring and heart disease}
\format{
Data frame with 8 observations on the following 3 variables
\describe{
  \item{\code{pat}}{presence of heart disease, factor with 2 levels (\code{no, si})}
  \item{\code{russ}}{level of snoring, factor with 4 levels (\code{mai}, no snoring;
  \code{a volte}, occasional snoring; \code{spesso}, snoring nearly every night;
  \code{sempre}, alwayssnoring;)}
  \item{\code{freq}}{frequency observed in the cross classification of the previous 2 variables}
}
}
\source{
Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J., Ostrowski, E. (1994).
  \emph{Small Data Sets}. London Chapman and Hall/CRC.
}
\usage{
Snore
}
\description{
Data from a report of a survey which investigated whether snoring was related to heart disease.
Those surveyed were classified according to the amount they snored, on the basis of reports
from their spouses.
}
\keyword{datasets}
