% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Spending}
\alias{Spending}
\title{Opinions about government spending}
\format{
A data frame with 81 observations on the following 5 variables
\describe{
  \item{\code{e}}{opinion on spending on the environment (\code{1}, too little; \code{2}, about right;
  \code{3}, too much)}
  \item{\code{h}}{opinion on spending on the health (\code{1}, too little; \code{2}, about right;
  \code{3}, too much)}
  \item{\code{c}}{opinion on spending on assistance to big cities (\code{1}, too little; \code{2}, about right;
  \code{3}, too much)}
  \item{\code{l}}{opinion on spending on law enforcement (\code{1}, too little; \code{2}, about right;
  \code{3}, too much)}
  \item{\code{count}}{frequency of subjects in the cross classification of the previous 4 variables}
}
}
\source{
Agresti, A. (2013). \emph{Categorical Data Analysis}, 3rd ed. New York: Wiley.

  \url{http://users.stat.ufl.edu/~aa/cda/data.html}
}
\usage{
Spending
}
\description{
Subjects in a 1989 General Social Survey  from the National Opinion Research Center in the U.S. were
asked their opinions about government spending on the environment (e), health (h),
assistance to big cities (c), and law enforcement (l).
}
\keyword{datasets}
