% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Stroke1}
\alias{Stroke1}
\title{Stroke data}
\format{
A data frame with 192 observations on the following 4 variables

\describe{
\item{\code{Subject}}{subject indicator}

\item{\code{Group}}{group indicator, factor with 3 levels (\code{A, B, C})}

\item{\code{Week}}{week indicator}

\item{\code{y}}{Bartel index}
}
}
\source{
Dobson, A. J. and Barnett A. (2008). \emph{An Introduction to Generalized Linear Models}, 3-rd ed.
Boca Raton: CRC Press.
}
\usage{
Stroke1
}
\description{
Longitudinal data from an experiment to promote the recovery of stroke patients in wide
format. The response variable is the Bartel index with higher scores meaning better
outcomes and a maximum score of 100.
}
\seealso{
\code{\link{Stroke}} for the same data in a different format
}
\keyword{datasets}
