% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Testingresso}
\alias{Testingresso}
\title{University admission test}
\format{
A data frame with 630 observations on the following 3 variables
\describe{
  \item{\code{y}}{indicator variable of correct answer (\code{1}, correct; \code{0}, wrong)}
  \item{\code{subject}}{candidate indicator (\code{1}, \ldots, \code{63})}
  \item{\code{item}}{question indicator (\code{1}, \ldots, \code{10})}
}
}
\source{
Salvan, A., Sartori, N., Pace, L. (2020). \emph{Modelli lineari generalizzati}. Milano: Springer-Verlag.
}
\usage{
Testingresso
}
\description{
Admission test for Statistical Sciences bachelor course at University of Padova in 2014/15.
The data refers to the answers of 63 candidates to 10 questions on text comprehension.
}
\keyword{datasets}
