% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Vehicle}
\alias{Vehicle}
\title{Preferred vehicle}
\format{
A data frame with 2067 observations on the following 4 variables
\describe{
  \item{\code{age}}{age of the owner}
  \item{\code{men}}{gender (\code{1}, man; \code{0}, female)}
  \item{\code{urban}}{residential area (\code{1}, urban; \code{0}, rural)}
  \item{\code{veh}}{kind of vehicle, factor with 3 levels (\code{C}, car; \code{F}, fourwheel; \code{M}, motorcycle)}
}
}
\source{
\url{http://www.ub.edu/rfa/R/regression_with_categorical_dependent_variables.html}

  Guillén, M. (2014). Regression with categorical dependent variables. In \emph{Predictive
  Modeling Applications in Actuarial Science - Volume I: Predictive Modeling Techniques}, E.W. Frees, R.A.
  Derrig and G. Meyers (Eds.) pp. 65-86. Cambridge: Cambridge University Press.
}
\usage{
Vehicle
}
\description{
Data from an insurance company, which record for each contract the kind of vehicle, together with
some additional variables.
}
\keyword{datasets}
