% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wool}
\alias{Wool}
\title{Wool data}
\format{
Data frame with 27 observations on the following 4 variables
\describe{
  \item{\code{x1}}{length of test specimen (\code{-1}, 250 mm; \code{0}, 300 mm; \code{1}, 350 mm)}
  \item{\code{x2}}{amplitude of loading cycle  (\code{-1}, 8 mm; \code{0}, 9 mm; \code{1}, 10 mm)}
  \item{\code{x3}}{load  (\code{-1}, 40 g; \code{0}, 45 g; \code{1}, 50 g)}
  \item{\code{y}}{cycles to failure}
}
}
\source{
Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J., Ostrowski, E. (1994).
  \emph{Small Data Sets}. London Chapman and Hall/CRC.
}
\usage{
Wool
}
\description{
The data show the number of cycles to failure of samples of worsted yarn under cycles of repeated loading.
There are three experimental conditions arranged in a 3 x 3 x 3 factorial design.
}
\keyword{datasets}
